refactor - #100 request로 표기된 값 response로 변경
