--- src/main/java/sopt/org/hmh/global/auth/social/kakao/request/KakaoTokenResponse.java
@@ -1,7 +1,0 @@
-package sopt.org.hmh.global.auth.social.kakao.request;
-
-public record KakaoTokenResponse(
-        String accessToken,
-        String refreshToken
-) {
-}
