+++ src/main/java/sopt/org/hmh/global/auth/social/kakao/response/KakaoAccount.java
@@ -1,11 +1,10 @@
-package sopt.org.hmh.global.auth.social.kakao.request;
+package sopt.org.hmh.global.auth.social.kakao.response;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-public record KakaoUserProfile(
-        String nickname,
-        String profileImageUrl
+public record KakaoAccount(
+        KakaoUserProfile profile
 ) {
 }
