+++ src/main/java/sopt/org/hmh/global/auth/social/kakao/response/KakaoUserProfile.java
@@ -1,11 +1,11 @@
-package sopt.org.hmh.global.auth.social.kakao.request;
+package sopt.org.hmh.global.auth.social.kakao.response;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-public record KakaoUserRequest(
-        Long id,
-        KakaoAccount kakaoAccount
+public record KakaoUserProfile(
+        String nickname,
+        String profileImageUrl
 ) {
 }
