+++ src/main/java/sopt/org/hmh/global/auth/social/kakao/response/KakaoUserResponse.java
@@ -1,10 +1,11 @@
-package sopt.org.hmh.global.auth.social.kakao.request;
+package sopt.org.hmh.global.auth.social.kakao.response;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-public record KakaoAccount(
-        KakaoUserProfile profile
+public record KakaoUserResponse(
+        Long id,
+        KakaoAccount kakaoAccount
 ) {
 }
